Improved the way batch sampling is done.
I removed the short-lived run_batch method and the batch_size constructor parameter. Whether to run in single or batch mode is now determined when run() is called, by calling it with either one or two arguments. If no batch size is provided the program runs in single mode.

The user can now choose to save the Markov chain to disk, in numpy format for the moment. In batch mode the saving can be done either to a single file, which is overwritten after each batch, or in separate files for each batch.